refactor(ranking): 오류 수정
- incrementScore에 event 시각의 hour를 전달
- order 신호 적재를 BigDecimal로 통일
- 미사용 incrementScoreSequential 데드코드 삭제
- cursor pagination이 deleted product로 조기 종료되지 않도록 수정
- API date/hour 입력 검증을 400으로 매핑
- ranking enrichment 실패 시 warn 로깅 추가
- ViewEventConsumer 디시리얼라이즈 에러 메시지 정정

+++ apps/commerce-streamer/src/main/java/com/loopers/domain/ranking/ProductDailySignalRepository.java
@@ -1,5 +1,6 @@
 package com.loopers.domain.ranking;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -9,7 +10,7 @@ public interface ProductDailySignalRepository {
 
     void upsertLikeCount(Long productDbId, LocalDate date, long delta);
 
-    void upsertOrderAmount(Long productDbId, LocalDate date, double amount);
+    void upsertOrderAmount(Long productDbId, LocalDate date, BigDecimal amount);
 
     List<ProductDailySignalModel> findBySignalDate(LocalDate date);
 }
